+++ Lombiq.Hosting.Tenants.EmailQuotaManagement/Services/QuotaService.cs
@@ -1,13 +1,22 @@
 ﻿using Lombiq.Hosting.Tenants.EmailQuotaManagement.Indexes;
 using Lombiq.Hosting.Tenants.EmailQuotaManagement.Models;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Options;
 using OrchardCore.Email;
 using OrchardCore.Environment.Shell.Configuration;
 using OrchardCore.Modules;
+using OrchardCore.Security;
+using OrchardCore.Security.Services;
+using OrchardCore.Users;
+using OrchardCore.Users.Models;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using YesSql;
+using static OrchardCore.Security.Permissions.Permission;
+using static OrchardCore.Security.StandardPermissions;
 
 namespace Lombiq.Hosting.Tenants.EmailQuotaManagement.Services;
 
@@ -18,19 +27,42 @@ public class QuotaService : IQuotaService
     private readonly IShellConfiguration _shellConfiguration;
     private readonly SmtpSettings _smtpOptions;
     private readonly IClock _clock;
+    private readonly IRoleService _roleService;
+    private readonly UserManager<IUser> _userManager;
 
     public QuotaService(
         ISession session,
         IOptions<EmailQuotaOptions> emailQuotaOptions,
         IShellConfiguration shellConfiguration,
         IOptions<SmtpSettings> smtpOptions,
-        IClock clock)
+        IClock clock,
+        IRoleService roleService,
+        UserManager<IUser> userManager)
     {
         _session = session;
         _emailQuotaOptions = emailQuotaOptions.Value;
         _shellConfiguration = shellConfiguration;
         _smtpOptions = smtpOptions.Value;
         _clock = clock;
+        _roleService = roleService;
+        _userManager = userManager;
+    }
+
+    public async Task<IEnumerable<string>> CollectUserEmailsForExceedingQuotaAsync()
+    {
+        // Get users with site owner permission.
+        var roles = await _roleService.GetRolesAsync();
+        var siteOwnerRoles = roles.Where(role =>
+            (role as Role)?.RoleClaims.Exists(claim =>
+                claim.ClaimType == ClaimType && claim.ClaimValue == SiteOwner.Name) == true);
+
+        var siteOwners = new List<IUser>();
+        foreach (var role in siteOwnerRoles)
+        {
+            siteOwners.AddRange(await _userManager.GetUsersInRoleAsync(role.RoleName));
+        }
+
+        return siteOwners.Select(user => (user as User)?.Email);
     }
 
     public bool ShouldLimitEmails()
@@ -93,15 +125,9 @@ public class QuotaService : IQuotaService
             return true;
         }
 
-        switch (emailQuota.LastReminderPercentage)
-        {
-            case >= 80 when currentPercentage < 90:
-            case >= 90 when currentPercentage < 100:
-            case >= 100:
-                return false;
-            default:
-                return true;
-        }
+        return !((emailQuota.LastReminderPercentage >= 80 && currentPercentage < 90) ||
+            (emailQuota.LastReminderPercentage >= 90 && currentPercentage < 100) ||
+            emailQuota.LastReminderPercentage >= 100);
     }
 
     public void ResetQuota(EmailQuota emailQuota)
